requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary
Werkzeug==2.3.7
email-validator==1.3.1
python-dotenv==1.0.0
Flask-Mail